# openapi.json
{"openapi":"3.1.0","info":{"title":"News API","description":"FastAPI + Meilisearchを使用した高速ニュース記事管理API","version":"1.0.0"},"paths":{"/api/v1/news/s3/health":{"get":{"tags":["news"],"summary":"Check S3 Health","description":"S3サービスのヘルスチェック","operationId":"check_s3_health_api_v1_news_s3_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/news/thumbnails/s3":{"post":{"tags":["news"],"summary":"Upload Thumbnail To S3","description":"S3にサムネイル画像をアップロードします","operationId":"upload_thumbnail_to_s3_api_v1_news_thumbnails_s3_post","parameters":[{"name":"cache_duration","in":"query","required":false,"schema":{"type":"string","description":"キャッシュ期間: short, medium, long","default":"long","title":"Cache Duration"},"description":"キャッシュ期間: short, medium, long"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_thumbnail_to_s3_api_v1_news_thumbnails_s3_post"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/S3UploadResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/thumbnails/s3/list":{"get":{"tags":["news"],"summary":"List S3 Thumbnails","description":"S3バケット内のサムネイル一覧を取得","operationId":"list_s3_thumbnails_api_v1_news_thumbnails_s3_list_get","parameters":[{"name":"prefix","in":"query","required":false,"schema":{"type":"string","description":"検索プレフィックス","default":"thumbnails/","title":"Prefix"},"description":"検索プレフィックス"},{"name":"max_keys","in":"query","required":false,"schema":{"type":"integer","description":"最大取得件数","default":100,"title":"Max Keys"},"description":"最大取得件数"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/thumbnails/s3/{filename}":{"delete":{"tags":["news"],"summary":"Delete S3 Thumbnail","description":"S3からサムネイル画像を削除します","operationId":"delete_s3_thumbnail_api_v1_news_thumbnails_s3__filename__delete","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/thumbnails":{"post":{"tags":["news"],"summary":"Upload Thumbnail Deprecated","description":"ローカルサムネイルアップロード（廃止済み）","operationId":"upload_thumbnail_deprecated_api_v1_news_thumbnails_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_thumbnail_deprecated_api_v1_news_thumbnails_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/thumbnails/{filename}":{"delete":{"tags":["news"],"summary":"Delete Thumbnail Deprecated","description":"ローカルサムネイル削除（廃止済み）","operationId":"delete_thumbnail_deprecated_api_v1_news_thumbnails__filename__delete","parameters":[{"name":"filename","in":"path","required":true,"schema":{"type":"string","title":"Filename"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news":{"post":{"tags":["news"],"summary":"Create Article","description":"新しい記事を作成します","operationId":"create_article_api_v1_news_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsArticleCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsArticle"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["news"],"summary":"Read Articles","description":"記事一覧を取得します","operationId":"read_articles_api_v1_news_get","parameters":[{"name":"skip","in":"query","required":false,"schema":{"type":"integer","default":0,"title":"Skip"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","default":10,"title":"Limit"}},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"}},{"name":"published","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Published"}},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"description":"タグでフィルタリング（複数指定可能）","title":"Tags"},"description":"タグでフィルタリング（複数指定可能）"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/facets":{"get":{"tags":["news"],"summary":"Get Facets","description":"ファセットカウントを取得します（カテゴリ、タグ、公開状態ごとの記事数）","operationId":"get_facets_api_v1_news_facets_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"検索クエリ（任意）","title":"Q"},"description":"検索クエリ（任意）"},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"カテゴリでフィルタリング","title":"Category"},"description":"カテゴリでフィルタリング"},{"name":"published","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"公開状態でフィルタリング","title":"Published"},"description":"公開状態でフィルタリング"},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"description":"タグでフィルタリング（複数指定可能）","title":"Tags"},"description":"タグでフィルタリング（複数指定可能）"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FacetResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/search":{"get":{"tags":["news"],"summary":"Search Articles Endpoint","description":"記事を検索します（検索クエリなしでフィルタリングのみも可能）","operationId":"search_articles_endpoint_api_v1_news_search_get","parameters":[{"name":"q","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"検索クエリ（任意）","title":"Q"},"description":"検索クエリ（任意）"},{"name":"category","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"カテゴリでフィルタリング","title":"Category"},"description":"カテゴリでフィルタリング"},{"name":"published","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"公開状態でフィルタリング","title":"Published"},"description":"公開状態でフィルタリング"},{"name":"tags","in":"query","required":false,"schema":{"anyOf":[{"type":"array","items":{"type":"string"}},{"type":"null"}],"description":"タグでフィルタリング（複数指定可能）","title":"Tags"},"description":"タグでフィルタリング（複数指定可能）"},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","description":"取得件数","default":10,"title":"Limit"},"description":"取得件数"},{"name":"offset","in":"query","required":false,"schema":{"type":"integer","description":"スキップ件数","default":0,"title":"Offset"},"description":"スキップ件数"},{"name":"sort_by","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"ソート順（例：created_at:desc）","title":"Sort By"},"description":"ソート順（例：created_at:desc）"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/news/{article_id}":{"get":{"tags":["news"],"summary":"Read Article","description":"指定されたIDの記事を取得します","operationId":"read_article_api_v1_news__article_id__get","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsArticle"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["news"],"summary":"Update Article","description":"指定されたIDの記事を更新します","operationId":"update_article_api_v1_news__article_id__put","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsArticleUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewsArticle"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["news"],"summary":"Delete Article","description":"指定されたIDの記事を削除します","operationId":"delete_article_api_v1_news__article_id__delete","parameters":[{"name":"article_id","in":"path","required":true,"schema":{"type":"integer","title":"Article Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contact":{"post":{"tags":["contact"],"summary":"Submit Contact Form","description":"お問い合わせフォームを送信します（SNS統合版）\n\n- **name**: お名前（必須）\n- **email**: メールアドレス（必須）\n- **subject**: 件名（必須）\n- **message**: お問い合わせ内容（必須）\n- **phone**: 電話番号（任意）\n- **company**: 会社名（任意）\n\n機能:\n- S3への保存\n- SNS通知（本番環境のみ）\n- 自動返信メール\n- レート制限","operationId":"submit_contact_form_api_v1_contact_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactForm"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contact/legacy":{"post":{"tags":["contact"],"summary":"Submit Contact Form Legacy","description":"お問い合わせフォームを送信します（従来版・後方互換性のため）\n\n- **name**: お名前（必須）\n- **email**: メールアドレス（必須）\n- **subject**: 件名（必須）\n- **message**: お問い合わせ内容（必須）\n- **phone**: 電話番号（任意）\n- **company**: 会社名（任意）","operationId":"submit_contact_form_legacy_api_v1_contact_legacy_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactForm"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/contact/sync":{"post":{"tags":["contact"],"summary":"Submit Contact Form Sync","description":"お問い合わせフォームを同期的に送信します（テスト用・SNS統合版）\n\n開発・テスト環境でメール送信の結果を即座に確認したい場合に使用","operationId":"submit_contact_form_sync_api_v1_contact_sync_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactForm"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/email/health":{"get":{"tags":["contact"],"summary":"Check Email Health","description":"メールサービスのヘルスチェック\n\nメール機能が正常に動作するかを確認します","operationId":"check_email_health_api_v1_email_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EmailHealthResponse"}}}}}}},"/api/v1/email/test":{"post":{"tags":["contact"],"summary":"Send Test Email","description":"テストメールを送信します（開発用）\n\nメール機能の動作確認用エンドポイント","operationId":"send_test_email_api_v1_email_test_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestEmailRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TestEmailResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Read Root","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_upload_thumbnail_deprecated_api_v1_news_thumbnails_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_thumbnail_deprecated_api_v1_news_thumbnails_post"},"Body_upload_thumbnail_to_s3_api_v1_news_thumbnails_s3_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_thumbnail_to_s3_api_v1_news_thumbnails_s3_post"},"ContactForm":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"subject":{"type":"string","title":"Subject"},"message":{"type":"string","title":"Message"},"phone":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Phone"},"company":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company"}},"type":"object","required":["name","email","subject","message"],"title":"ContactForm","description":"お問い合わせフォームのスキーマ"},"ContactResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"},"contact_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Contact Id"},"errors":{"anyOf":[{"items":{},"type":"array"},{"type":"null"}],"title":"Errors"}},"type":"object","required":["success","message"],"title":"ContactResponse","description":"お問い合わせレスポンス"},"EmailHealthResponse":{"properties":{"status":{"type":"string","title":"Status"},"is_local":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Local"},"mail_server":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Mail Server"},"mail_port":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Mail Port"},"template_folder":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Template Folder"},"services":{"anyOf":[{"additionalProperties":{"type":"boolean"},"type":"object"},{"type":"null"}],"title":"Services"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error"}},"type":"object","required":["status"],"title":"EmailHealthResponse","description":"メールサービスヘルスチェックレスポンス"},"FacetCount":{"properties":{"value":{"type":"string","title":"Value"},"count":{"type":"integer","title":"Count"}},"type":"object","required":["value","count"],"title":"FacetCount","description":"ファセットカウントの結果"},"FacetResponse":{"properties":{"categories":{"items":{"$ref":"#/components/schemas/FacetCount"},"type":"array","title":"Categories","default":[]},"tags":{"items":{"$ref":"#/components/schemas/FacetCount"},"type":"array","title":"Tags","default":[]},"published":{"items":{"$ref":"#/components/schemas/FacetCount"},"type":"array","title":"Published","default":[]},"total_articles":{"type":"integer","title":"Total Articles","default":0}},"type":"object","title":"FacetResponse","description":"ファセットカウントのレスポンス"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"NewsArticle":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"published":{"type":"boolean","title":"Published","default":false},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"thumbnail_alt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Alt"},"created_at":{"type":"string","format":"date-time","title":"Created At"},"updated_at":{"type":"string","format":"date-time","title":"Updated At"}},"type":"object","required":["id","title","content","created_at","updated_at"],"title":"NewsArticle"},"NewsArticleCreate":{"properties":{"title":{"type":"string","title":"Title"},"content":{"type":"string","title":"Content"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"tags":{"items":{"type":"string"},"type":"array","title":"Tags","default":[]},"published":{"type":"boolean","title":"Published","default":false},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"thumbnail_alt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Alt"}},"type":"object","required":["title","content"],"title":"NewsArticleCreate"},"NewsArticleUpdate":{"properties":{"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"},"category":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Category"},"author":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Author"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags"},"published":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Published"},"thumbnail_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Url"},"thumbnail_alt":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Thumbnail Alt"}},"type":"object","title":"NewsArticleUpdate"},"S3UploadResponse":{"properties":{"s3_url":{"type":"string","title":"S3 Url"},"cloudfront_url":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Cloudfront Url"},"filename":{"type":"string","title":"Filename"},"size":{"type":"integer","title":"Size"},"cache_control":{"type":"string","title":"Cache Control"}},"type":"object","required":["s3_url","filename","size","cache_control"],"title":"S3UploadResponse","description":"S3アップロードのレスポンス"},"SearchResponse":{"properties":{"items":{"items":{"$ref":"#/components/schemas/NewsArticle"},"type":"array","title":"Items"},"total":{"type":"integer","title":"Total"},"limit":{"type":"integer","title":"Limit"},"offset":{"type":"integer","title":"Offset"}},"type":"object","required":["items","total","limit","offset"],"title":"SearchResponse"},"TestEmailRequest":{"properties":{"to_email":{"type":"string","format":"email","title":"To Email"},"subject":{"type":"string","title":"Subject","default":"テストメール"},"message":{"type":"string","title":"Message","default":"これはテストメールです。"}},"type":"object","required":["to_email"],"title":"TestEmailRequest","description":"テストメール送信リクエスト"},"TestEmailResponse":{"properties":{"success":{"type":"boolean","title":"Success"},"message":{"type":"string","title":"Message"}},"type":"object","required":["success","message"],"title":"TestEmailResponse","description":"テストメール送信レスポンス"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}